# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.AddFavoriteRequest:
    properties:
      station_id:
        type: string
    required:
    - station_id
    type: object
...
  internal_handlers.ErrorDetail:
    properties:
      code:
        example: STATION_NOT_FOUND
        type: string
      field:
        example: email (only present on validation errors)
        type: string
      message:
        example: station not found
        type: string
    type: object
  internal_handlers.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handlers.ErrorDetail'
    type: object
  internal_handlers.FavoriteListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_handlers.FavoriteStationDTO'
        type: array
      meta:
        $ref: '#/definitions/internal_handlers.MetaCount'
    type: object
  internal_handlers.FavoriteStationDTO:
    properties:
      country:
        example: United States
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      image_url:
        example: https://cdn.rockfm.com/logo.png
        type: string
      is_premium_only:
        example: false
        type: boolean
      name:
        example: Rock FM 100.1
        type: string
      stream_url:
        example: https://stream.rockfm.com/live
        type: string
      tags:
        example:
        - rock
        - classic rock
        items:
          type: string
        type: array
      votes:
        example: 1500
        type: integer
    type: object
...
  internal_handlers.GuestRateLimitData:
    properties:
      enabled:
        example: false
        type: boolean
      limit_per_hour:
        example: 100
        type: integer
    type: object
  internal_handlers.GuestRateLimitResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handlers.GuestRateLimitData'
      success:
        example: true
        type: boolean
    type: object
...
  internal_handlers.MetaCount:
    properties:
      count:
        example: 20
        type: integer
    type: object
...
  internal_handlers.SimpleErrorResponse:
    properties:
      error:
        example: authentication required
        type: string
    type: object
...
  internal_handlers.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@radiobackend.com
    name: API Support
  description: "Radio streaming API with JWT authentication, analytics and multi-language
    support (i18n).\n\n**Security Features (v2.1):**\n- \U0001F510 Timing attack prevention
    (constant-time operations)\n- \U0001F512 HTTPS enforcement (production)\n- \U0001F6AB
    Account lockout (10 attempts = 30 min lockout)\n- \U0001F4E7 Email rate limiting
    (10 attempts/hour per email)\n- \U0001F511 Password strength enforcement (special
    chars + 47 common passwords blocked)\n- \U0001F6E1️ Session hijacking protection
    (User-Agent validation)\n- \U0001F4DD Security event logging (failed attempts
    with metadata)\n- ⚡ Rate limiting (5 attempts/15min per IP)\n- \U0001F510 JWT
    with RS256 + token revocation\n- \U0001F916 Bot protection via shared secret (X-Rradio-Secret
    header)\n\n**Internationalization:**\nSupported languages: Spanish (es), English
    (en), French (fr), German (de).\nUse the 'lang' parameter or 'Accept-Language'
    header to specify the desired language.\n\n**Security Score:** 100/100 (audited
    Dec 2025)"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Radio Backend API
  version: "2.1"
paths:
...
  /api/v1/admin/security/guest-rate-limit:
    get:
      consumes:
      - application/json
      description: Returns whether the per-guest-IP rate limiter is currently enabled
        and the configured limit.
      produces:
      - application/json
      responses:
        "200":
          description: Current status
          schema:
            $ref: '#/definitions/internal_handlers.GuestRateLimitResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
      security:
      - BearerAuth: []
      summary: Get guest IP rate limit status
      tags:
      - Admin Security
  /api/v1/admin/security/guest-rate-limit/toggle:
    post:
      consumes:
      - application/json
      description: Enables the limiter if it is currently disabled, and vice versa.
        Returns the new state.
      produces:
      - application/json
      responses:
        "200":
          description: New state
          schema:
            $ref: '#/definitions/internal_handlers.GuestRateLimitResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Failed to toggle guest rate limiter
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle guest IP rate limiter
      tags:
      - Admin Security
...
  /favorites:
    get:
      consumes:
      - application/json
      description: Returns the list of favorite stations for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of favorite stations
          schema:
            $ref: '#/definitions/internal_handlers.FavoriteListResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user favorites
      tags:
      - Favorites
    post:
      consumes:
      - application/json
      description: Adds a station to the authenticated user's favorites list
      parameters:
      - description: Station ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.AddFavoriteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Favorite added successfully
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "403":
          description: Access denied - Premium only station
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Station not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "409":
          description: Station already in favorites
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add station to favorites
      tags:
      - Favorites
  /favorites/{stationId}:
    delete:
      consumes:
      - application/json
      description: Removes a station from the authenticated user's favorites list
      parameters:
      - description: Station ID
        in: path
        name: stationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Favorite removed successfully
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Station ID is required
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "404":
          description: Favorite not found
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove station from favorites
      tags:
      - Favorites
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  SharedSecret:
    description: Shared secret key required for public station and SEO endpoints (bot
      protection).
    in: header
    name: X-Rradio-Secret
    type: apiKey
swagger: "2.0"